:wrench: add configResults function in Context
