--- src/components/Admin.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-const Admin = () => {
-    return (
-        <h1>
-            Admin
-        </h1>
-    )
-}
-export default Admin
