--- src/components/User.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-const User = () => {
-    return (
-        <h1>
-            User
-        </h1>
-    )
-}
-export default User
